feat: moved helper functions to 'pypi' class
