+++ src/data_provider/src/packages/packageLinuxParserDeb.cpp
@@ -12,8 +12,6 @@
 #include "sharedDefs.h"
 #include "packageLinuxParserHelper.h"
 #include <fstream>
-#include <iostream>
-#include <unordered_set>
 
 void getDpkgInfo(const std::string& fileName, std::function<void(nlohmann::json&)> callback)
 {
@@ -48,58 +46,5 @@ void getDpkgInfo(const std::string& fileName, std::function<void(nlohmann::json&
                 callback(packageInfo);
             }
         }
-    }
-}
-
-void getDpkgPythonPackages(std::unordered_set<std::string>& pythonPackages)
-{
-    std::regex listPattern(R"(^python.*\.list$)");
-    const auto PYTHON_INFO_FILES = std::array {std::make_pair(std::regex(R"(^.*\.egg-info$)"), "/PKG-INFO"),
-                                               std::make_pair(std::regex(R"(^.*\.dist-info$)"), "/METADATA")};
-
-    try
-    {
-        for (const auto& entry : std::filesystem::directory_iterator(DPKG_INFO_PATH))
-        {
-            if (std::filesystem::is_regular_file(entry) &&
-                    std::regex_search(entry.path().filename().string(), listPattern))
-            {
-                std::ifstream file(entry.path());
-                std::string line;
-
-                while (std::getline(file, line))
-                {
-                    std::smatch match;
-
-                    for (const auto& [pattern, extraFile] : PYTHON_INFO_FILES)
-                    {
-                        if (std::regex_search(line, match, pattern))
-                        {
-                            std::string baseInfoPath = match.str(0);
-
-                            if (std::filesystem::is_regular_file(baseInfoPath))
-                            {
-                                pythonPackages.insert(baseInfoPath);
-                            }
-                            else
-                            {
-                                std::string fullPath = baseInfoPath + extraFile;
-
-                                if (std::filesystem::exists(fullPath))
-                                {
-                                    pythonPackages.insert(fullPath);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                file.close();
-            }
-        }
-    }
-    catch (const std::filesystem::filesystem_error& ex)
-    {
-        std::cerr << "Filesystem error: " << ex.what() << std::endl;
     }
 }
